+++ lib/tools/pencil_controller.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:sketcher/tools/tool_controller.dart';
+import 'package:sketcher/ui/operations/stroke_operation.dart';
+import 'package:sketcher/ui/reactive_painter.dart';
 import 'package:sketcher/ui/sketch_controller.dart';
 
 class PencilController implements ToolController {
@@ -9,21 +11,30 @@ class PencilController implements ToolController {
   PencilController(this._sketchController, this._onStateUpdated);
 
   @override
+  ReactivePainter toolPainter;
+
+  @override
   void panStart(PointerDownEvent details) {
-    _sketchController.reactivePainter.startStroke(details.localPosition);
+    toolPainter = ReactivePainter(_sketchController.activeToolStyle);
+    toolPainter.startStroke(details.localPosition);
+    _onStateUpdated();
   }
 
   @override
   void panUpdate(PointerMoveEvent details) {
-    _sketchController.reactivePainter.appendStroke(details.localPosition);
+    toolPainter.appendStroke(details.localPosition);
   }
 
   @override
   void panEnd(PointerUpEvent details) {
-    _sketchController.reactivePainter.endStroke();
-    if (_sketchController.reactivePainter.strokes.length > 10) {
-      _sketchController.commitStrokes();
-      _onStateUpdated();
-    }
+    final stroke = toolPainter.endStroke();
+    toolPainter = null;
+    final operation = StrokeOperation(stroke);
+    _sketchController.commitOperation(operation);
+    // _sketchController.reactivePainter.endStroke();
+    // if (_sketchController.reactivePainter.strokes.length > 10) {
+    //   _sketchController.commitStrokes();
+    //   _onStateUpdated();
+    // }
   }
 }