+++ lib/ui/operations/operation.dart
@@ -1,0 +1,7 @@
+import 'package:sketcher/sketcher.dart';
+
+abstract class Operation {
+  void redo(SketchController controller);
+
+  void undo(SketchController controller);
+}