+++ lib/ui/sketch_layer.dart
@@ -1,0 +1,8 @@
+import 'package:sketcher/ui/static_painter.dart';
+
+class SketchLayer {
+  final int id;
+  final StaticPainter painter;
+
+  SketchLayer(this.id, this.painter);
+}