+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-## [1.0.0]
+## 1.0.1
+
+* added the undo/redo actions for eraser
+
+## 1.0.0
 
 * added the pencil tool
 * added the eraser tool
-* added the undo/redo actions
+* added the undo/redo actions for pencil
 * added the SVG converter